requirements.txt
Django==3.0.3
django-filter==2.2.0
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
wincertstore==0.2
